Create switchcase
how does the switch case work in the if else construct in c++
View other drafts






















The switch case statement in C++ can be used within an if-else construct to implement more complex logic. For example, the following code uses a switch case statement to select the appropriate action to take based on the value of a variable:

C++
int choice = 1;

if (choice >= 1 && choice <= 3) {
  switch (choice) {
    case 1:
      std::cout << "You chose option 1." << std::endl;
      break;
    case 2:
      std::cout << "You chose option 2." << std::endl;
      break;
    case 3:
      std::cout << "You chose option 3." << std::endl;
      break;
  }
} else {
  std::cout << "Invalid choice." << std::endl;
}
Use code with caution. Learn more
In this example, the choice variable is first checked to make sure that it is between 1 and 3. If it is, then the switch case statement is used to select the appropriate action to take. If the choice variable is not between 1 and 3, then the default code block is executed.

Another common use case for the switch case statement within an if-else construct is to handle errors. For example, the following code uses a switch case statement to handle different types of errors that may occur when reading a file:

C++
std::ifstream file("myfile.txt");

if (file.is_open()) {
  // Read the file and handle any errors
  switch (file.rdstate()) {
    case std::ifstream::eofbit:
      // End of file reached
      std::cout << "End of file reached." << std::endl;
      break;
    case std::ifstream::failbit:
      // An error occurred while reading the file
      std::cout << "An error occurred while reading the file." << std::endl;
      break;
    case std::ifstream::badbit:
      // A fatal error occurred while reading the file
      std::cout << "A fatal error occurred while reading the file." << std::endl;
      break;
  }

  file.close();
} else {
  std::cout << "Could not open file." << std::endl;
}
Use code with caution. Learn more
In this example, the file.rdstate() member function is used to check the state of the file stream. If the state is eofbit, then the end of the file has been reached. If the state is failbit, then an error occurred while reading the file. If the state is badbit, then a fatal error occurred while reading the file.

The switch case statement is a powerful tool that can be used to implement complex logic in C++. By using the switch case statement within an if-else construct, you can write code that is more robust and easier to maintain.