Create switch with for loop
o print the string "*****\n" five times in C++, you can use the following code:

C++
#include <iostream>

using namespace std;

int main() {
  for (int i = 0; i < 5; i++) {
    cout << "(*****\n)";
  }
  return 0;
}
Use code with caution. Learn more
This code will print the string "*****\n" to the console five times. The for loop iterates from 0 to 4, and on each iteration, the string is printed to the console using the cout statement.

Here is an example of the output of the program:

(*****
(*****
(*****
(*****
(*****
You can also use a while loop to print the string five times. The following code is equivalent to the previous example:

C++
#include <iostream>

using namespace std;

int main() {
  int i = 0;
  while (i < 5) {
    cout << "(*****\n)";
    i++;
  }
  return 0;
}
Use code with caution. Learn more
This code will print the string "*****\n" to the console five times, and then it will terminate.

I hope this helps!