Create stack multiple cases
Here are the main things to know about the switch statement:

The expression must be an int, or int-compatible like char or enum (to be covered later in the course).
The alignment of the nested keywords case: are often not indented from the contining switch, as one might have expected. switch statement indentation usually follows the above alignment. (However, the code under each case: should be indented.)
You cannot have ranges in cases. There is no (sensible) way to cover "an int between 3 and 200."
The break; statements are not required, but they enform specific behavior. If a break; is missing from a case: block, control leaks down to the next case:. This is sometimes done intentionally. But if a break; is included (as above), contol passes to the the statement following the entire switch{ } block.
You can stack multiple cases for the same code block. (You'll see this in the next example.)
Here an example that uses some of these details: