Create switchFunction
Switch Case Example: Switcheroo
In this example, we will throw out the if elseif statements and use the switch case statement instead. Switch case statements are a substitute for long if statements that compare a variable to several "integral" values.  "Integral" values are simply values that can be expressed as an integer, such as the value of a int or char.  In other words, you can't use switch case statement with strings in c++--just integers or values that can be interpreted as integers.  We will use integers in this example. 

We will create a program that displays a menu for the user.   This menu gives the user the option of running one of three class functions: playgame, loadgame or playmultiplayer.  Each of these functions has a different message to display. 

Note: This program is a modified version of the program presented in the Tutorial: switch case.  Our version compares the if/else construct to the switch/case construct.

First, let's implement the three functions that will be called. These functions each display a simple message using cout. The implementation generally goes right beneath the class definition in the header file.

void playgame()
{
        cout << "Play game called\n";
}
void loadgame()
{
	cout << "Load game called\n";
}
void playmultiplayer()
{
	cout << "Play multiplayer game called\n";
}
We need to display the menu and put in the conditions for when each of the above functions will be called. The functions setData and getFunction will do this for us.  First, let's see how this would look using if/else, then we will rewrite the code using switch case.

void setData()
{
  //declare variable called input here
  cout<<"1. Play game\n";
  cout<<"2. Load game\n";
  cout<<"3. Play multiplayer\n";
  cout<<"4. Exit\n";
  cout<<"Selection: ";
  cin>> input;
  //TO DO: call getFunction here
  //or call switchFunction here
}

void getFunction(int input)
{
  if(input == 1) 
      playgame();
  else if(input==2)
      loadgame();
  else if(input==3)
      playmultiplayer();
  else if(input==4)
      cout<<"Thank you for playing!\n";
  else
      cout<<"Error, bad input, quitting\n";
}

void switchFunction(int input)
{
  switch ( input ) 
  {
    case 1:            // Note the colon, not a semicolon
      playgame();
      break;
    case 2:            // Note the colon, not a semicolon
      loadgame();
      break;
    case 3:            // Note the colon, not a semicolon
      playmultiplayer();
      break;
    case 4:            // Note the colon, not a semicolon
      cout<<"Thank you for playing!\n";
      break;
    default:            // Note the colon, not a semicolon
      cout<<"Error, bad input, quitting\n";
      break;
  }
}

In the .cpp file we need to write the main function. The main function only needs to call the function setData.