Create ENUMS
This program uses an enumerated type called wind_directions_t to represent the different wind directions. The enumerated type has five possible values: NORTH_WIND, SOUTH_WIND, EAST_WIND, WEST_WIND, and NO_WIND.

The program first prompts the user to enter a wind direction. The user's input is then converted to an enumerated value using a series of if statements. If the user's input is not a valid wind direction, the program prints an error message and exits.

Once the wind direction has been converted to an enumerated value, the program uses a switch statement to print the appropriate message to the console.

Here is an example of the program's output:

Please enter NORTH, SOUTH, EAST, WEST, or NONE for our wind direction
NORTH

The wind is blowing from the north.