Create ENUM.SWITCH.ELSE.IF
what does return 1; mean in c++
In C++, return 1; means that the function in which it is used is terminating abnormally and is returning a failure code to the calling function.

The value 1 is a common failure code, but other values can also be used. It is up to the programmer to decide what values to use to indicate different types of failures.

For example, a function that reads a file might return 1 if the file cannot be opened. A function that writes to a database might return 1 if the database connection cannot be established.

The calling function can then check the return value to determine whether the function was successful. If the return value is 1, or any other non-zero value, the calling function knows that the function failed and can take appropriate action.

For example, the following code checks the return value of the open() function to determine whether the file was opened successfully:

C++
int main() {
  int fd = open("myfile.txt", O_RDONLY);
  if (fd == -1) {
    // File could not be opened.
    return 1;
  }

  // File was opened successfully.
  // ...

  close(fd);
  return 0;
}
Use code with caution. Learn more
It is important to note that return 1; is not only used in user-defined functions. It can also be used in the main() function to indicate that the program terminated abnormally.

For example, the following code returns 1 from the main() function to indicate that there was an error:

C++
int main() {
  // ...

  // There was an error.
  return 1;
}
Use code with caution. Learn more
In general, it is good practice to use return values to indicate the success or failure of functions. This makes code more robust and easier to maintain.

profile picture
what does the switch statement mean in c++
View other drafts























1

2
The switch statement in C++ is a flow control statement that allows you to execute different blocks of code based on the value of an expression. It is an alternative to the long if-else-if ladder, and it can be used to simplify code and make it more readable.

The syntax of the switch statement is as follows:

C++
switch (expression) {
  case value1:
    // Code to be executed if expression == value1
    break;
  case value2:
    // Code to be executed if expression == value2
    break;
  ...
  default:
    // Code to be executed if no case matches
}
Use code with caution. Learn more
The expression is evaluated once, and the value is compared to each of the case values. If there is a match, the code associated with that case is executed. If there is no match, the code associated with the default case is executed.

The break statement is optional, but it is good practice to use it. It prevents the code from executing the cases that follow the matched case.

For example, the following code uses a switch statement to determine the day of the week based on the value of a variable called dayOfWeek:

C++
int main() {
  int dayOfWeek = 1; // Monday

  switch (dayOfWeek) {
    case 1:
      std::cout << "Today is Monday." << std::endl;
      break;
    case 2:
      std::cout << "Today is Tuesday." << std::endl;
      break;
    case 3:
      std::cout << "Today is Wednesday." << std::endl;
      break;
    ...
    default:
      std::cout << "Invalid day of week." << std::endl;
  }

  return 0;
}
Use code with caution. Learn more
Output:

Today is Monday.
Switch statements can be used to implement a variety of different algorithms. For example, they can be used to implement state machines, routing tables, and menus.

Here are some tips for using switch statements effectively:

Use break statements to prevent the code from executing the cases that follow the matched case.
Use default cases to handle unexpected values.
Use nested switch statements to implement complex logic.
Document your switch statements carefully so that other programmers can understand them easily.
Sources
github.com/HyunCafe/my-javascript-notes
github.com/ALMSALMA/Cpp20_tutorial