Create arrays
what kind of output would this code produce.....for (k = 0; k < 20; k++)
      student[k] = 7;
View other drafts















The code snippet for (k = 0; k < 20; k++) { student[k] = 7; } will produce an array student with 20 elements, all of which are set to the value 7.

The for loop iterates from 0 to 19, and on each iteration, the value of k is used to index into the array student. The assignment operator = is then used to set the value of the element at index k to 7.

Here is an example of the output:

student = [7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7]
This code snippet could be used to initialize an array of student grades, for example.