Create intSIZEintARRAY
what constitutes integer size in the array  int array[] = {5, 3, 4, 1, 2, 33,345,122,111222,444,333333,352525,5435345,345454,435545435,34553646,45654656};
 int size = sizeof(array) / sizeof(array[0]);
View other drafts




















The integer size in the array int array[] = {5, 3, 4, 1, 2, 33,345,122,111222,444,333333,352525,5435345,345454,435545435,34553646,45654656}; is the size of the int type on the system where the code is compiled.

The size of the int type is typically 4 bytes on most modern systems. This means that each element in the array will use 4 bytes of memory.

The variable size is calculated using the following formula:

size = sizeof(array) / sizeof(array[0]);
The sizeof() operator returns the size of the data type in bytes. So, the formula above calculates the total size of the array in bytes, and then divides it by the size of one element in the array to get the number of elements in the array.

In this case, the variable size will be equal to 17, because there are 17 elements in the array.

Here is an example of how to use the sizeof() operator to calculate the size of an array in C:

C
#include <stdio.h>

int main() {
  int array[] = {5, 3, 4, 1, 2};
  int size = sizeof(array) / sizeof(array[0]);

  printf("The size of the array is %d\n", size);

  return 0;
}
Use code with caution. Learn more
Output:

"
The size of the array is 5
"